types

+++ trunk/APPS/incast/elephant_sink.c
@@ -45,7 +45,7 @@ process_a_child(int sd, struct sockaddr_in *sin, int use_sctp)
 	if ((no_clock_e == 0) && (no_clock_s == 0)) {
 		timespecsub(&tve, &tvs);
 		printf("Read %ld times for %ld bytes in %ld.%9.9ld from ",
-		       cnt, sz, (long int)tve.tv_sec, 
+		       (long int)cnt, (long int)sz, (long int)tve.tv_sec, 
 		       tve.tv_nsec);
 		print_an_address((struct sockaddr *)sin, 1);
 	}
